+++ py/desisurvey/ephemerides.py
@@ -169,6 +169,12 @@ class Ephemerides(object):
         # tabulating the (ra,dec) of objects to avoid.
         t_obj = np.linspace(0., 1., num_obj_steps)
 
+        # Lookup sun altitude thresholds.
+        bright_max_alt = (
+            config.programs.BRIGHT.max_sun_altitude().to(u.rad).value)
+        dark_max_alt = (
+            config.programs.DARK.max_sun_altitude().to(u.rad).value)
+
         # Calculate ephmerides for each night.
         for day_offset in range(num_nights):
             day = self.start + day_offset * u.day
@@ -177,15 +183,13 @@ class Ephemerides(object):
             # Store local noon for this day.
             row['noon'] = day.mjd
             # Calculate bright twilight.
-            mayall.horizon = (
-                config.programs.BRIGHT.max_sun_altitude().to(u.rad).value)
+            mayall.horizon = bright_max_alt
             row['brightdusk'] = mayall.next_setting(
                 ephem.Sun(), use_center=True) + mjd0
             row['brightdawn'] = mayall.next_rising(
                 ephem.Sun(), use_center=True) + mjd0
             # Calculate dark / gray twilight.
-            mayall.horizon = (
-                config.programs.DARK.max_sun_altitude().to(u.rad).value)
+            mayall.horizon = dark_max_alt
             row['dusk'] = mayall.next_setting(
                 ephem.Sun(), use_center=True) + mjd0
             row['dawn'] = mayall.next_rising(
@@ -291,8 +295,7 @@ class Ephemerides(object):
         Parameters
         ----------
         mjd : float or array
-            MJD values during a single night where the program should be
-            tabulated.
+            MJD values where the illuminated fraction should be tabulated.
 
         Returns
         -------
@@ -421,6 +424,60 @@ class Ephemerides(object):
         sort_order = np.argsort(cycle)
         # Return True if tonight's illumination is in the top num_nights.
         return index - lo in sort_order[-num_nights:]
+
+
+def get_sun_altitude(row, mjd):
+    """Return the sun altitude at the specified time for twilight modeling.
+
+    This function only returns values in the range [-10, -20] deg
+    since values outside this range are either never used for observing
+    or else considered completely dark.
+
+    The calculation uses linear interpolation of the tabulated times when the
+    sun passes through the bright and dark maximum altitudes, nominally -13
+    and -15 deg, which provides an altitude accurate to better than 30"
+    between these limits.
+
+    Parameters
+    ----------
+    row : astropy.table.Row
+        A single row from the ephemerides astropy Table corresponding to the
+        night in question.
+    mjd : float
+        MJD value(s) during a single night where the sun altitude should be
+        calculated.
+
+    Returns
+    -------
+    astropy.units.Quantity
+        Sun altitude(s) in degrees at each input mjd value, clipped to
+        the range [-10, -20].
+    """
+    # Lookup the bright, dark threshold altitudes.
+    config = desisurvey.config.Configuration()
+    bright_max_alt = config.programs.BRIGHT.max_sun_altitude()
+    dark_max_alt = config.programs.DARK.max_sun_altitude()
+
+    # Validate input timestamp(s)
+    mjd = np.asarray(mjd)
+    scalar = mjd.ndim == 0
+    mjd = np.atleast_1d(mjd)
+    midnight = 0.5 * (row['dusk'] + row['dawn'])
+    if np.any(np.abs(mjd - midnight) > 0.5):
+        raise ValueError('mjd is not associated with the specified night')
+
+    # Use linear interpolation on the dusk data for times before midnight.
+    sun_alt = np.empty_like(mjd) * u.degree
+    at_dusk = mjd < midnight
+    sun_alt[at_dusk] = bright_max_alt + (dark_max_alt - bright_max_alt) * (
+        mjd[at_dusk] - row['brightdusk']) / (row['dusk'] - row['brightdusk'])
+
+    # Use linear interpolation on the dawn data for times after midnight.
+    at_dawn = ~at_dusk
+    sun_alt[at_dawn] = bright_max_alt + (bright_max_alt - dark_max_alt) * (
+        mjd[at_dawn] - row['brightdawn']) / (row['brightdawn'] - row['dawn'])
+
+    return np.clip(sun_alt, -20 * u.deg, -10 * u.deg)
 
 
 def get_object_interpolator(row, object_name, altaz=False):