+++ doc/changes.rst
@@ -13,6 +13,7 @@ desisurvey change log
 0.10.1 (2017-12-20)
 -------------------
 
+* Add Ephemerides method to calculate sun altitude during twilight.
 * Set the ``EXTNAME`` keyword on the Table returned by ``Progress.get_exposures()``.
 
 0.10.0 (2017-11-09)